fix: validate DeFi permit token address OK-56972
